refactor: centralize the Article-related APIs
