+++ src/articles/articles.module.ts
@@ -2,6 +2,7 @@ import { HttpModule } from "@nestjs/axios";
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { IpfsModule } from "src/ipfs/ipfs.module";
+import { UsersModule } from "src/users/users.module";
 
 import { ArticlesController } from "./articles.controller";
 import { ArticlesService } from "./articles.service";
@@ -13,6 +14,7 @@ import { Comment } from "./entities/comment.entity";
   imports: [
     HttpModule,
     IpfsModule,
+    UsersModule,
     TypeOrmModule.forFeature([Article, Comment]),
     CommentsModule,
   ],
