+++ src/users/users.module.ts
@@ -1,6 +1,5 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { ArticlesModule } from "src/articles/articles.module";
 import { MailModule } from "src/mail/mail.module";
 
 import { User } from "./entities/user.entity";
@@ -8,7 +7,7 @@ import { UsersController } from "./users.controller";
 import { UsersService } from "./users.service";
 
 @Module({
-  imports: [MailModule, ArticlesModule, TypeOrmModule.forFeature([User])],
+  imports: [MailModule, TypeOrmModule.forFeature([User])],
   controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
